refactor: #9 Remote Layer mapper 분리
