+++ remote/src/main/java/com/kdjj/remote/datasource/RecipeRemoteDataSourceImpl.kt
@@ -2,8 +2,8 @@ package com.kdjj.remote.datasource
 
 import com.kdjj.data.datasource.RecipeRemoteDataSource
 import com.kdjj.domain.model.Recipe
-import com.kdjj.remote.dto.toDomain
-import com.kdjj.remote.dto.toDto
+import com.kdjj.remote.mapper.toDomain
+import com.kdjj.remote.mapper.toDto
 import com.kdjj.remote.service.RemoteRecipeService
 import javax.inject.Inject
