+++ remote/src/main/java/com/kdjj/remote/dto/RecipeStepDto.kt
@@ -1,6 +1,5 @@
 package com.kdjj.remote.dto
 
-import com.kdjj.domain.model.RecipeStep
 import com.kdjj.domain.model.RecipeStepType
 
 internal data class RecipeStepDto(
@@ -11,23 +10,3 @@ internal data class RecipeStepDto(
     val imgPath: String = "",
     val seconds: Int = 0
 )
-
-internal fun RecipeStepDto.toDomain(): RecipeStep =
-    RecipeStep(
-        stepId,
-        name,
-        type,
-        description,
-        imgPath,
-        seconds
-    )
-
-internal fun RecipeStep.toDto(): RecipeStepDto =
-    RecipeStepDto(
-        stepId,
-        name,
-        type,
-        description,
-        imgPath,
-        seconds
-    )
