+++ remote/src/main/java/com/kdjj/remote/dto/RecipeTypeDto.kt
@@ -1,20 +1,6 @@
 package com.kdjj.remote.dto
 
-import com.kdjj.domain.model.RecipeType
-
 internal data class RecipeTypeDto(
     val id: Int = 0,
     val title: String = ""
 )
-
-internal fun RecipeTypeDto.toDomain(): RecipeType =
-    RecipeType(
-        id,
-        title
-    )
-
-internal fun RecipeType.toDto(): RecipeTypeDto =
-    RecipeTypeDto(
-        id,
-        title
-    )
