+++ remote/src/main/java/com/kdjj/remote/mapper/RecipeTypeDtoMapper.kt
@@ -1,0 +1,16 @@
+package com.kdjj.remote.mapper
+
+import com.kdjj.domain.model.RecipeType
+import com.kdjj.remote.dto.RecipeTypeDto
+
+internal fun RecipeTypeDto.toDomain(): RecipeType =
+    RecipeType(
+        id,
+        title
+    )
+
+internal fun RecipeType.toDto(): RecipeTypeDto =
+    RecipeTypeDto(
+        id,
+        title
+    )
